API: Make context manager base class for ptychofft
Because:
- context managers handle interrupts automatically
- context managers are guarantee calls to __exit__
- separating the base ptycho operators makes code modular

NOTE: GPU memory is no longer freed when the process is stopped because
stopped processed can be resumed, so freeing memory would cause resumed
processes to error.

The Solver class has been broken into parent and child classes:
PtychoCuFFT and CGPtychoSolver. The parent class contains only the
ptychography kernels such as fwd and adj, and the child class contains
only the congjuate gradient algorithm functions. This makes it easier to
implmement other solver algorithms without needing to copy code or make
the Solver class super large.

Additionally, the base class is now a context manager in order to make
GPU memory management easier and to automatically clean up the memory
when SIGINT is sent.