REF: Make PtychoCuFFT a parent of ptychofft
Because:
- Using inheritance reduces duplicate attributes

By having PtychoCuFFT inherit from ptychofft, all of the size variables
are not redefined in the python layer.

In addition, free() was added to the C++ layer such that PtychoCuFFT can
free the memory of its parent when __exit__ is called. I haven't figured
out how to overload __enter__ and __exit__ from the C++ layer because
these functions accept or return Python types that do not have obvious C
equivalents.