API: Return results in a dictionary
Because:
- Returning a dictionary makes future API changes backwards compatible
- Solvers may return different results without breaking clients
- Dictionaries are self-documenting because keywords

When using a dictionary, we only catch one result, then use a keyword
to extract the information we want. This is similar to returning a
struct in C/C++. By returning a dictionary instead of a tuple, we can
add or remove variables from the result arbitrarily. For example, when
the API changed from solving only for psi to solving both psi and prb,
an additional variable was returned, and all client code now needed
to catch two variables. In the future, adding return variables such as
alignment information(?) will not break old client code.